+++ fedify/federation/queue.ts
@@ -3,6 +3,15 @@ export interface SenderKeyJwkPair {
   privateKey: JsonWebKey;
 }
 
+/**
+ * A message that represents a task to be processed by the background worker.
+ * The concrete type of the message depends on the `type` property.
+ *
+ * Please do not depend on the concrete types of the messages, as they may
+ * change in the future.  You should treat the `Message` type as an opaque
+ * type.
+ * @since 1.6.0
+ */
 export type Message = FanoutMessage | OutboxMessage | InboxMessage;
 
 export interface FanoutMessage {